pyproject.toml
[project]
name = "rolex-search"
version = "0.1.0"
description = ""
authors = []
dependencies = [
    "openai>=1.0.0",
    "pandas>=1.0.0",
    "openpyxl>=3.0.0",
    "python-dotenv>=0.20.0", # .envファイル読み込み用
    "rich>=13.0.0",          # コンソール出力整形用
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
include = ["/main.py"]
